Move VSep and HSep to the widgets module
